Fix data returned in setFriendlyDescription graphql mutation
closes OUT-4891
flag=none

Test plan:
- Create outcome with description
and non-empty friendly description.
- Open dev tools and go to the networks tab
- Edit the friendly description so it
is now empty.
- Find the graphql event near the bottom where the
operation names equals "SetOutcomeFriendlyDescription".
- Verify the description field in both the payload column
and response column are both "".

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/325772
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Angela Gomba <angela.gomba@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>

+++ app/graphql/mutations/set_friendly_description.rb
@@ -57,17 +57,15 @@ class Mutations::SetFriendlyDescription < Mutations::BaseMutation
       friendly_description.workflow_state = "active"
       friendly_description.description = description
       friendly_description.save!
-      return {
-        outcome_friendly_description: friendly_description
-      }
-    elsif friendly_description.persisted?
-      friendly_description.destroy
-      return {
-        outcome_friendly_description: friendly_description
-      }
+
+    else
+      friendly_description.destroy if friendly_description.persisted?
+      friendly_description.description = ""
     end
 
-    {}
+    {
+      outcome_friendly_description: friendly_description
+    }
   end
 
   private
